+++ apps/monitor-web/src/components/common/Buttons/index.ts
@@ -1,0 +1,3 @@
+export { default as BasicButton } from "./BasicButton";
+export { default as ClearButton } from "./ClearButton";
+export { default as IconButton } from "./IconButton";
