+++ apps/monitor-web/src/components/common/feedback/LoadingSpinner.tsx
@@ -1,4 +1,4 @@
-import Icon from "./Icon";
+import Icon from "../Icon";
 import { cn } from "@/utils";
 
 /**
