+++ apps/monitor-web/src/components/common/index.ts
@@ -1,14 +1,8 @@
+export * from "./Buttons";
+export * from "./feedback";
+export * from "./inputs";
+
 export { default as Checkbox } from "./Checkbox";
-export { default as ClearButton } from "./ClearButton";
-export { default as ErrorBoundary } from "./ErrorBoundary";
 export { default as Icon } from "./Icon";
 export type { IconName } from "./Icon";
-export { default as LoadingSpinner } from "./LoadingSpinner";
-export { default as Nav } from "./Nav";
-export { default as TextareaField } from "./TextareaField";
-export { default as SearchInput } from "./SearchInput";
-export { default as TextField } from "./TextField";
-export { default as Toast } from "./Toast";
 export { default as Image } from "./Image";
-export { default as BasicButton } from "./Buttons/BasicButton";
-export { default as IconButton } from "./Buttons/IconButton";
