+++ apps/monitor-web/src/components/common/inputs/index.ts
@@ -1,0 +1,3 @@
+export { default as SearchInput } from "./SearchInput";
+export { default as TextField } from "./TextField";
+export { default as TextareaField } from "./TextareaField";
