+++ apps/monitor-web/src/layouts/index.ts
@@ -1,1 +1,2 @@
+export { default as Nav } from "./Nav";
 export { default as ToastContainer } from "./ToastContainer";
